Move Space class to core.py
